src/chainlit-app.py
import base64
import os
from typing import Dict, Optional

import chainlit as cl
from chainlit.types import ThreadDict
from openai import AsyncOpenAI
from jwt.exceptions import InvalidTokenError
from fastapi import Request, Response
import http
import jwt

from src.services.promptservice import get_system
from src.services.adminservice import get_user_from_identifier
from src.db.database import get_db
from src.core.config import Settings
from src.models.auth import TokenData
from src.schema.models import UserRole
from src.services.ragutils import ingestion
from src.services.ragutils import retrieval
from datetime import datetime, timezone


VLLM_BASE_URL = os.getenv("VLLM_BASE_URL", "http://localhost:8000")
client = AsyncOpenAI(base_url=f"{VLLM_BASE_URL}/v1", api_key="empty")
cl.instrument_openai()
SYSTEM_PROMPT = get_system()
settings = {"model": "Qwen/Qwen3-VL-8B-Instruct", "temperature": 0.7}


@cl.on_chat_resume
async def on_chat_resume(thread: ThreadDict):
    cl.user_session.set("message_history", [])

    for message in thread["steps"]:
        if message["type"] == "user_message":
            cl.user_session.get("message_history").append(
                {"role": "user", "content": message["output"]}
            )
        elif message["type"] == "assistant_message":
            cl.user_session.get("message_history").append(
                {"role": "assistant", "content": message["output"]}
            )


@cl.header_auth_callback
def header_auth_callback(headers: Dict) -> Optional[cl.User]:
    token = headers.get("bearer")

    if not token:
        return None

    app_settings = Settings()

    try:
        db_generator = get_db()
        db = next(db_generator)
        payload = jwt.decode(
            token, app_settings.AUTH_SECRET, algorithms=[app_settings.HASH_ALGORITHM]
        )
        username = payload.get("sub")
        expire_at = payload.get("exp")
        if username is None:
            return None
        elif datetime.now(timezone.utc) > datetime.fromtimestamp(
            expire_at, timezone.utc
        ):
            return None
        user = get_user_from_identifier(identifier=username, db=db)
    except InvalidTokenError:
        return None
    except Exception as e:
        return None

    if user is None:
        return None
    elif user.role == UserRole.unauthorized:
        return None

    return cl.User(
        identifier=f"{user.identifier}",
        metadata={"role": f"{user.role}", "provider": "header"},
    )


@cl.on_logout
def logout(request: Request, response: Response):
    response.delete_cookie("my_cookie")


@cl.on_chat_start
def on_start():
    cl.user_session.set(
        "message_history",
        [{"content": f"{SYSTEM_PROMPT}", "role": "system"}],
    )


@cl.on_message
async def on_message(cl_msg: cl.Message):

    user = cl.user_session.get("user")
    user_id = user.identifier if user else "anonymous"

    IMAGES = []
    DOCS = []

    for file in cl_msg.elements:
        if "image" in file.mime:
            IMAGES.append(file)
        else:
            DOCS.append(file)

    if DOCS:
        for doc in DOCS:
            count, error = ingestion.ingest_file(
                file_path=doc.path,
                file_id=doc.id,
                file_name=doc.name,
                file_type=doc.mime,
                user_id=user_id,
            )
            if error:
                await cl.Message(content=f"⚠️ {error}").send()
            elif count > 0:
                await cl.Message(content=f"✅ Ingested **{doc.name}** ({count} chunks)").send()

    context_str, sources = retrieval.get_context(cl_msg.content, user_id)

    final_query = cl_msg.content
    if context_str:
        final_query = f"Document Context: {context_str}\n\nUser Query: {cl_msg.content}"

    message = {"role": "user", "content": [{"type": "text", "text": final_query}]}

    for image in IMAGES:
        with open(image.path, "rb") as image_file:
            b64_image = base64.b64encode(image_file.read()).decode("utf-8")
        message["content"].append(
            {
                "type": "image_url",
                "image_url": {"url": f"data:image/png;base64,{b64_image}"},
            },
        )

    message_history = cl.user_session.get("message_history")
    message_history.append(message)

    msg = cl.Message(content="")

    stream = await client.chat.completions.create(
        messages=message_history, stream=True, **settings
    )

    async for part in stream:
        if token := part.choices[0].delta.content or "":
            await msg.stream_token(token)

    message_history.append({"role": "assistant", "content": msg.content})
    await msg.send()
