src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.30.0",
    "boto3>=1.42.39",
    "chainlit==2.9.6",
    "chromadb>=1.5.1",
    "fastapi[standard]>=0.121.3",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-milvus>=0.3.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.7.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pypdf>=6.7.3",
    "scrypt>=0.9.4",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "deepeval>=3.7.2",
    "groq>=0.36.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "pytest-sugar>=1.1.1",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rA -vv --tb=short"
testpaths = ["../tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "live: marks tests that require the running vLLM container",
]
